Add GPU processing packages and supporting software
GPUmatrix needs a python installation with tensorflow
or the R torch package.

Python with tensorflow must be set up as part of the
system environment, I couldn't get it to work as a
build input to the R package. The python package
reticulate needs to know the python path, and I can only
pass it in as an environmental variable.

There might be smarter hacks, like modifying reticulate,
but it could also break other setups that want reticulate
to work with other python packages too.

The R package torch needs to compile it's own software.
I found it easier to patch up the binaries provided, and
use my own branch of nixpkgs.